Add files via upload
Some ideas about how to start the pipeline. Starts by asking the user for fastq files then analyzes the file(s) with FastQC to determine the quality of the reads. Not sure how to properly load the all the modules that a user would actually need, so that part needs work
possible future directions - based on the quality of the reads could either send the files to a program like TRIMMOMATIC to clean up the file (if the sequences are of high enough quality would simply skip this step - would ideally use something like a while statement to do this). After this sequence could be aligned, converted to a SAM file, converted to a BAM file, the use platypus for variant calling.