Update Pipeline.sh
So changed how we run sickle to a for loop that will take any files in the raw/fastq_files directory that has a .fastq ending and will run the sabre program on all of them and should place them in raw/fastq_trimmed/ directory with the ending .fq for the next step.
At the moment this only works for single end reads, not paired end. working on a way to identify if the reads are paired end or single and then a way to add an option to the loop to process each with the correct sickle argument.