Update Pipeline.sh
so changed the name of some of the directories to make it less confusing where things are coming from. Added a line at the bottom of the for loop that counts read coverage. there are a couple more lines we could add to this that could take us all the way to variant calling with samtools but it requires a tool within samtools but i don't know how to call it if we are using a variable to use the tool, also doesn't really matter if we are going to use platypus anyway. What we could do here is break up this loop into smaller loops so we can add check points so if you have different file formats, like a sam file, you could jump to that part of the pipeline and start from there